Move restartloop.sh back to root
This is the script that actually runs the "daemon" in the foreground.
Having it in the base directory is useful, because that will typically
be the working directory.

Maybe at some point we can turn it into a nice systemd unit. But for
now, screen with a shell loop will do :-)